@@ -71,7 +71,7 @@
     "template.data.gouv.fr": "^1.3.2",
     "tmp": "^0.2.1",
     "validator": "^13.7.0",
-    "vue": "^2.6.14",
+    "vue": "^3.2.22",
     "vue-async-computed": "^3.9.0",
     "vue-matomo": "^4.1.0",
     "vue-router": "^3.5.3",
