@@ -8,7 +8,8 @@ module.exports = {
     "plugin:vue/essential",
     "prettier",
     "plugin:cypress/recommended",
-    "plugin:prettier-vue/recommended",
+    //"plugin:prettier-vue/recommended",
+    "plugin:vue/vue3-recommended",
   ],
   plugins: ["prettier", "vue"],
   globals: {
